refactor: Break down massive App.tsx into modular components
- Reduce App.tsx from 1,005 lines to 24 lines (96% reduction)
- Extract components into separate files for better maintainability:
  * AppContent.tsx (456 lines) - Main application logic and layout
  * Canvas.tsx (232 lines) - Canvas workspace with drag-and-drop
  * CanvasElement.tsx (55 lines) - Individual canvas element rendering
  * ControlPanel.tsx (51 lines) - Properties panel component
  * ElementLibrary.tsx (183 lines) - Component library with loading
- Create organized type definitions:
  * AppTypes.ts (32 lines) - App-level interfaces and types
  * ThemeProvider.tsx (38 lines) - Theme management provider
- Fix import paths and type compatibility issues
- Maintain full MCO/MSO architecture functionality across all components
- Preserve React StrictMode compatibility and proper cleanup
- Enable better code organization for team development

BREAKING CHANGE: App.tsx now uses modular component architecture
All functionality preserved with improved maintainability and testability

Benefits: Single responsibility principle, better tree-shaking, isolated testing,
improved developer experience, and easier maintenance
