+++ testdata/RenderX/src/providers/ThemeProvider.tsx
@@ -1,0 +1,38 @@
+/**
+ * Theme Provider Component
+ * Handles dynamic theme loading from Theme.theme-management-symphony plugin
+ */
+
+import React from 'react';
+
+// Import ThemeProvider from Theme Management Symphony Plugin
+// Note: In production, this would be dynamically loaded by the plugin system
+// For now, we'll use a placeholder that gets replaced by the plugin
+let ThemeProvider = ({ children }: { children: React.ReactNode }) => (
+  <>{children}</>
+);
+
+// This will be replaced by the plugin system when Theme.theme-management-symphony loads
+if (
+  typeof window !== "undefined" &&
+  (window as any).renderxPlugins?.["Theme.theme-management-symphony"]
+) {
+  ThemeProvider = (window as any).renderxPlugins[
+    "Theme.theme-management-symphony"
+  ].ThemeProvider;
+}
+
+// ThemeToggleButton now provided by Theme.theme-management-symphony plugin
+let ThemeToggleButton = () => <button>🎨</button>; // Fallback
+
+// This will be replaced by the plugin system when Theme.theme-management-symphony loads
+if (
+  typeof window !== "undefined" &&
+  (window as any).renderxPlugins?.["Theme.theme-management-symphony"]
+) {
+  ThemeToggleButton = (window as any).renderxPlugins[
+    "Theme.theme-management-symphony"
+  ].ThemeToggleButton;
+}
+
+export { ThemeProvider, ThemeToggleButton };
